requirements.txt
Django==1.9
psycopg2
markdown
matplotlib
decorator==4.0.2
django-guardian
django-crispy-forms
django-extensions
djangorestframework
django-filter
django-taggit
django-taggit-templatetags
django-form-utils
numpy
python-social-auth==0.2.7
requests
requests-oauthlib
jsonfield
django-dbbackup<2.3
dropbox==1.6
kombu
opbeat
djrill
ua-parser
user-agents
django-user-agents
django-dirtyfields
cython
h5py
Pillow
python-openid
django-sendfile
django-polymorphic
celery[redis]
django-celery
django-cleanup
django-chosen
git+https://github.com/sinnwerkstatt/django-file-resubmit.git#egg=file_resubmit
django-hstore==1.3.5
django-datatables-view
django-oauth-toolkit==0.9.0
wdl
django-gravatar2
singularity
simplejson
pygments
